chore: gitignore every course except "test" (#1829)
When someone develop a course it will show up in these two directories
and we would like to ignore them. We cannot gitignore the whole
directories /apps/web/src/courses/ and /courses/ because we have the
"test" directory in both.
We can assume that every course will have a dash in its name so
we ignore those.